+++ miniblog/app/main/routes.py
@@ -1,0 +1,141 @@
+from flask import render_template, flash, redirect, url_for, request, g, current_app
+from flask_login import current_user, login_required
+from flask_babel import _, get_locale
+from datetime import datetime, timezone
+import sqlalchemy as sa
+from app import db
+from app.main.forms import EditProfileForm, EmptyForm, PostForm
+from app.models import User, Post
+from app.translate import translate
+from langdetect import detect, LangDetectException
+from app.main import bp
+
+@bp.before_app_request
+def before_request():
+    if current_user.is_authenticated:
+        current_user.last_seen = datetime.now(timezone.utc)
+        db.session.commit()
+        g.locale = str(get_locale())
+
+@bp.route('/', methods=['GET', 'POST'])
+@bp.route('/index', methods=['GET', 'POST'])
+@login_required
+def index():
+    form = PostForm()
+    if form.validate_on_submit():
+        try:
+            language = detect(form.post.data)
+        except LangDetectException:
+            language = ''
+        post = Post(body=form.post.data, author=current_user,
+                    language=language)
+        db.session.add(post)
+        db.session.commit()
+        flash(_('你的貼文現在已發布！'))
+        return redirect(url_for('main.index'))
+    page = request.args.get('page', 1, type=int)
+    posts = db.paginate(current_user.following_posts(), page=page,
+                        per_page=current_app.config['POSTS_PER_PAGE'], error_out=False)
+    next_url = url_for('main.index', page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('main.index', page=posts.prev_num) \
+        if posts.has_prev else None
+    return render_template('index.html', title=_('首頁'), form=form,
+                           posts=posts.items, next_url=next_url,
+                           prev_url=prev_url)
+
+@bp.route('/user/<username>')
+@login_required
+def user(username):
+    user = db.first_or_404(sa.select(User).where(User.username == username))
+    page = request.args.get('page', 1, type=int)
+    query = user.posts.select().order_by(Post.timestamp.desc())
+    posts = db.paginate(query, page=page,
+                        per_page=current_app.config['POSTS_PER_PAGE'],
+                        error_out=False)
+    next_url = url_for('main.user', username=user.username, page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('main.user', username=user.username, page=posts.prev_num) \
+        if posts.has_prev else None
+    form = EmptyForm()
+    return render_template('user.html', user=user, posts=posts.items,
+                           next_url=next_url, prev_url=prev_url, form=form)
+
+@bp.route('/edit_profile', methods=['GET', 'POST'])
+@login_required
+def edit_profile():
+    form = EditProfileForm(current_user.username)
+    if form.validate_on_submit():
+        current_user.username = form.username.data
+        current_user.about_me = form.about_me.data
+        db.session.commit()
+        flash(_('你的修改已被儲存。'))
+        return redirect(url_for('main.edit_profile'))
+    elif request.method == 'GET':
+        form.username.data = current_user.username
+        form.about_me.data = current_user.about_me
+    return render_template('edit_profile.html', title=_('編輯個人資料'),
+                           form=form)
+
+@bp.route('/follow/<username>', methods=['POST'])
+@login_required
+def follow(username):
+    form = EmptyForm()
+    if form.validate_on_submit():
+        user = db.session.scalar(
+            sa.select(User).where(User.username == username))
+        if user is None:
+            flash(_('使用者 %(username)s 未找到。', username=username))
+            return redirect(url_for('main.index'))
+        if user == current_user:
+            flash(_('你不能追蹤自己！'))
+            return redirect(url_for('main.user', username=username))
+        current_user.follow(user)
+        db.session.commit()
+        flash(_('你正在追蹤 %(username)s ！', username=username))
+        return redirect(url_for('main.user', username=username))
+    else:
+        return redirect(url_for('main.index'))
+
+@bp.route('/unfollow/<username>', methods=['POST'])
+@login_required
+def unfollow(username):
+    form = EmptyForm()
+    if form.validate_on_submit():
+        user = db.session.scalar(
+            sa.select(User).where(User.username == username))
+        if user is None:
+            flash(_('使用者 %(username)s 未找到。', username=username))
+            return redirect(url_for('main.index'))
+        if user == current_user:
+            flash(_('你不能取消追蹤自己！'))
+            return redirect(url_for('main.user', username=username))
+        current_user.unfollow(user)
+        db.session.commit()
+        flash(_('你已經取消追蹤 %(username)s ！', username=username))
+        return redirect(url_for('main.user', username=username))
+    else:
+        return redirect(url_for('main.index'))
+
+@bp.route('/explore')
+@login_required
+def explore():
+    page = request.args.get('page', 1, type=int)
+    per_page = current_app.config['POSTS_PER_PAGE']
+    # post總數
+    posts_count = db.session.scalar(sa.text("SELECT COUNT(*) FROM post"))
+    query = sa.text(f"SELECT * FROM post ORDER BY timestamp DESC LIMIT {per_page} OFFSET {(page - 1) * per_page}")
+    posts = db.session.scalars(sa.select(Post).from_statement(query)).all()
+    # 前幾頁的資料數量 < post總數，代表後面還有post，下一頁開啟
+    next_url = url_for('main.explore', page=page + 1) if (page * per_page) < posts_count else None
+    prev_url = url_for('main.explore', page=page - 1) if page > 1 else None
+    return render_template("index.html", title=_('探索'), posts=posts,
+                           next_url=next_url, prev_url=prev_url)
+
+@bp.route('/translate', methods=['POST'])
+@login_required
+def translate_text():
+    data = request.get_json()
+    return {'text': translate(data['text'],
+                              data['source_language'],
+                              data['dest_language'])}
