+++ miniblog/app/models.py
@@ -1,13 +1,19 @@
+from flask import current_app
+from flask_login import UserMixin
 from datetime import datetime, timezone
 from typing import Optional
 import sqlalchemy as sa
 import sqlalchemy.orm as so
 from werkzeug.security import generate_password_hash, check_password_hash
-from flask_login import UserMixin
 from hashlib import md5
-from app import db, app, login
+from app import db, login
 from time import time
 import jwt
+
+@login.user_loader
+def load_user(id):
+    user = db.session.scalars(sa.select(User).from_statement(sa.text(f"SELECT * FROM user WHERE id = {id}"))).first()
+    return user
 
 followers = sa.Table(
     'followers',
@@ -91,12 +97,12 @@ class User(UserMixin, db.Model):
     def get_reset_password_token(self, expires_in=600):
         return jwt.encode(
             {'reset_password': self.id, 'exp': time() + expires_in},
-            app.config['SECRET_KEY'], algorithm='HS256')
+            current_app.config['SECRET_KEY'], algorithm='HS256')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
-            id = jwt.decode(token, app.config['SECRET_KEY'],
+            id = jwt.decode(token, current_app.config['SECRET_KEY'],
                             algorithms=['HS256'])['reset_password']
         except:
             return
@@ -108,8 +114,8 @@ class Post(db.Model):
     timestamp = sa.Column('timestamp', sa.DateTime, index=True, default=lambda: datetime.now(timezone.utc))
     user_id = sa.Column('user_id', sa.Integer, sa.ForeignKey("user.id"), index=True)
     author: so.Mapped[User] = so.relationship(back_populates='posts')
-    language = sa.Column('language', sa.String(5))
-
+    language: so.Mapped[Optional[str]] = so.mapped_column(sa.String(5))
+    
     def __repr__(self):
         return '<Post {}>'.format(self.body)
 
@@ -118,8 +124,5 @@ class Post(db.Model):
 # with app.app_context():
 #     db.create_all()
     
-@login.user_loader
-def load_user(id):
-    user = db.session.scalars(sa.select(User).from_statement(sa.text(f"SELECT * FROM user WHERE id = {id}"))).first()
-    return user
+
