+++ miniblog/app/templates/_post.html
@@ -1,18 +1,17 @@
 <table class="table table-hover">
   <tr>
     <td width="70px">
-      <a href="{{ url_for('user', username=post.author.username) }}">
+      <a href="{{ url_for('main.user', username=post.author.username) }}">
         <img src="{{ post.author.avatar(70) }}" />
       </a>
     </td>
     <td>
       {% set user_link %}
-      <a href="{{url_for('user', username=post.author.username) }}">
+      <a href="{{url_for('main.user', username=post.author.username) }}">
         {{post.author.username}}
       </a>
       {% endset %} {{_('%(username)s 說 %(when)s', username=user_link,
       when=moment(post.timestamp).fromNow()) }}
-
       <br />
       <span id="post{{ post.id}}">{{ post.body }}</span>
       {% if post.language and post.language != g.locale %}
