+++ miniblog/app/templates/auth/register.html
