+++ miniblog/app/templates/base.html
@@ -20,7 +20,7 @@
   <body>
     <nav class="navbar navbar-expand-lg bg-body-tertiary">
       <div class="container">
-        <a class="navbar-brand" href="{{ url_for('index') }}"
+        <a class="navbar-brand" href="{{ url_for('main.index') }}"
           >{{ _('微型部落格') }}</a
         >
         <button
@@ -40,7 +40,7 @@
               <a
                 class="nav-link"
                 aria-current="page"
-                href="{{ url_for('index') }}"
+                href="{{ url_for('main.index') }}"
                 >{{ _('首頁') }}</a
               >
             </li>
@@ -48,7 +48,7 @@
               <a
                 class="nav-link"
                 aria-current="page"
-                href="{{ url_for('explore') }}"
+                href="{{ url_for('main.explore') }}"
                 >{{ _('探索') }}</a
               >
             </li>
@@ -59,7 +59,7 @@
               <a
                 class="nav-link"
                 aria-current="page"
-                href="{{ url_for('login') }}"
+                href="{{ url_for('auth.login') }}"
                 >{{ _('登入') }}</a
               >
             </li>
@@ -68,7 +68,7 @@
               <a
                 class="nav-link"
                 aria-current="page"
-                href="{{ url_for('user', username=current_user.username) }}"
+                href="{{ url_for('main.user', username=current_user.username) }}"
                 >{{ _('個人資料') }}</a
               >
             </li>
@@ -76,7 +76,7 @@
               <a
                 class="nav-link"
                 aria-current="page"
-                href="{{ url_for('logout') }}"
+                href="{{ url_for('auth.logout') }}"
                 >{{ _('登出') }}</a
               >
             </li>
