+++ miniblog/app/templates/user.html
@@ -3,21 +3,22 @@
   <tr valign="top">
     <td><img src="{{ user.avatar(128) }}" /></td>
     <td>
-      <h1>使用者：{{ user.username }}</h1>
+      <h1>{{ _('使用者') }}：{{ user.username }}</h1>
+
       {% if user.about_me %}
       <p>{{ user.about_me }}</p>
       {% endif %} {% if user.last_seen %}
       <p>{{ _('最後上線時間') }}：{{ moment(user.last_seen).format('LLL') }}</p>
       {% endif %} {% if user == current_user %}
       <p>
-        <a href="{{ url_for('edit_profile') }}">
+        <a href="{{ url_for('main.edit_profile') }}">
           {{ _('編輯你的個人資料') }}
         </a>
       </p>
       {% elif not current_user.is_following(user) %}
       <div>
         <form
-          action="{{ url_for('follow', username=user.username) }}"
+          action="{{ url_for('main.follow', username=user.username) }}"
           method="post"
         >
           {{form.hidden_tag() }} {{form.submit(value=_('追蹤'), class='btn
@@ -27,7 +28,7 @@
       {% else %}
       <div>
         <form
-          action="{{ url_for('unfollow', username=user.username) }}"
+          action="{{ url_for('main.unfollow', username=user.username) }}"
           method="post"
         >
           {{form.hidden_tag() }} {{form.submit(value=_('取消追蹤'), class='btn
@@ -41,7 +42,7 @@
 <hr />
 {% for post in posts %} {% include '_post.html' %} {% endfor %}
 
-<nav aria-label="帖子導航">
+<nav aria-label="Post navigation">
   <ul class="pagination">
     <li class="page-item{% if not prev_url %} disabled{% endif %}">
       <a class="page-link" href="{{ prev_url}}">
